requirements.txt
asyncio
pymongo[srv]
telethon
python-dotenv
